pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.build]
include = ["src/**/*.py"]

[tool.hatch.build.targets.wheel.sources]
"src" = ""

[tool.isort]
profile = "black"
known_third_party = ["wandb"]
line_length = 100

[tool.ruff]
line-length = 100

[tool.docformatter]
black = true

[project]
name = "SmoothMyStrategy"
version = "0.0.0"
description = "Private package to have fun around finance."
authors = [{ name = "Yann POURCENOUX", email = "yann.pourcenoux@gmail.com" }]
readme = "README.md"
license = { file = "LICENSE" }
urls = { Homepage = "https://github.com/yann-pourcenoux/SmoothMyStrategy" }
requires-python = ">=3.12"
dependencies = [
    "altair>=5.5.0",
    "docformatter>=1.7.7",
    "GitPython>=3.1.44",
    "hydra-core>=1.3.2",
    "icecream>=2.1.5",
    "isort>=6.0.1",
    "loguru>=0.7.3",
    "pandas>=2.3.1",
    "polars>=1.31.0",
    "pre-commit>=4.2.0",
    "prettier>=0.0.7",
    "pydantic>=2.11.7",
    "pyportfolioopt>=1.5.6",
    "pytest>=4.6.11",
    "pytest-xdist>=1.34.0",
    "python-dotenv>=1.1.1",
    "pyupgrade>=3.20.0",
    "quantstats-lumi>=1.0.1",
    "ruff>=0.12.4",
    "seaborn>=0.13.2",
    "statsmodels>=0.14.5",
    "streamlit>=1.47.0",
    "tensordict>=0.9.1",
    "torch>=2.7.1",
    "torchrl>=0.9.2",
    "tqdm>=4.67.1",
    "wandb>=0.21.0",
    "yfinance>=0.2.65",
]

[project.scripts]
run_evaluation = "entrypoints:run_evaluation"
run_training = "entrypoints:run_training"
run_visualization = "entrypoints:run_visualization"
run_test = "entrypoints:run_test"
run_download = "entrypoints:run_download"
